requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
comm==0.2.2
constantly==23.10.4
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
daphne==4.1.0
debugpy==1.8.1
decorator==5.1.1
Django==5.0.3
django-filter==24.1
djangorestframework==3.15.0
executing==2.0.1
fonttools==4.50.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
Markdown==3.6
matplotlib==3.8.3
matplotlib-inline==0.1.6
mypy==1.9.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==25.1.2
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4
traitlets==5.14.2
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
wcwidth==0.2.13
zope.interface==6.2
